updated the test for dataloader; image keys handled
